refactor(direct-control): extract session execute wrappers
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves queryCiv7TunerStates, executeCiv7Command, executeCiv7AppUiCommand, and executeCiv7TunerCommand into src/session/execute.ts while keeping public facade exports and higher-level direct-control call sites stable.

This preserves session creation/close behavior, App UI/Tuner state forcing, command result shape, and focused session proof. index.ts still owns reconnect orchestration, lifecycle composition, public facade call-through for higher-level atoms, telemetry/correlation contracts, public procedure schemas, semantic CLI projection, AI ingestion, hotseat runtime proof, and Effect/oRPC procedure-core work. Task 2.9.4 matrix-row acceptance remains pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- session.test.ts public-api.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- git diff --cached --check
- ownership scan: Tuner query/execute wrappers are owned by src/session/execute.ts
- no-cycle scan: session modules do not import executable facade values

Local package/source relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
